[Refactor] Reduce position_t size.

+++ typedefs.h
@@ -98,7 +98,7 @@ struct pos_store_t {
     int epsq;
     int pliesFromNull;
     EvalScore posScore[2];
-    int mat_summ[2];
+    int16 mat_summ[2];
     uint64 phash;
     uint64 hash;
     pos_store_t* previous;
@@ -115,9 +115,9 @@ struct position_t {
     uint64 color[2];
     uint64 occupied;
     int8 pieces[64];
-    int kpos[2];
+    int8 kpos[2];
     pos_store_t posStore;
-    int side;
+    int8 side;
 };
 
 typedef uint8 mflag_t;
